cmd/internal/obj: write object file in new format
DO NOT REVIEW

If -newobj is set, write object file in new format, which uses
indices for symbol references instead of symbol names. The file
format is described at the beginning of
cmd/internal/obj/objfile2.go.

The object file writer is written in a way that trys to make the
encoding as regular as possible, and the reader and writer as
symmetric as possible.

This is incomplete, and currently nothing will consume the new
object file.

Change-Id: Ifefedbf6456d760d15a9f40a28af6486c93100fe