Update patch set 3
Patch Set 3:

In security.go line 192 there was already a regex for this option. I changed the unit-test to use this instead. 

I am on a linux machine, so I can't check the Darwin linker. But a test file with the following seems to work:

// #cgo LDFLAGS: -Wl,-sectcreate,__TEXT,__info_plist,${SRCDIR}/Info.plist -Wl,-sectcreate,__TEXT,__launchd_plist,${SRCDIR}/Launchd.plist

Patch-set: 3
Reviewer: Gerrit User 33592 <33592@62eb7196-b449-3ce5-99f1-c037f21e1705>