Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I don't have strong opinion. Just for the note. This make forcibly the filename encode with rtc 6266. So older browser will not work after this change. Do we need to leave a way to output raw-filenames?

I've filed https://github.com/golang/go/issues/33801 to keep track of these discussions and also for other evaluations.

Pascal, thank you for this CL and others for the review! When you make an update to the commit message, please add to the end of your commit message:

Fixes #33801.

Patch-set: 1
CC: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>