Update patch set 1
Patch Set 1:

Now that I think about it, this change will break the functionality of FindProcess() if somebody wants to use the handle returned with the Process.

My usecase was just to check if a process is still running, and notify if the process is terminated. But since, the handle is never closed, my program fails to notify on process exit.

Patch-set: 1
CC: Gerrit User 33648 <33648@62eb7196-b449-3ce5-99f1-c037f21e1705>