Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Robert, could you have another look at this for 1.14 when you have some time? Happy to try another approach if you think this isn't good enough.

Ok. I'm still not very comfortable with this as is. The parser uses "_" as identifier if something went wrong in the hope that "_" avoids future errors; e.g., it's always assignable to (but on the other hand it's not readable). "_" is also always ok in a declaration. An empty string can mean something different (and valid) depending on context (but maybe that's taken care of elsewhere). 

Using a special identifier ("?") may be better - but have you thought through the consequences of that? For instance, now, if one writes code where we expect an identifier to be declared and there's something else we get a syntax error and a "_" for it and the compiler will automatically do the right thing. For instance, with this code:

package p
var . int
var . int

the errors are correct: "syntax error: unexpected ., expecting name". If the scanner uses a "?" instead of a "_" in this case, the same code leads to an additional error: "? redeclared in this block" unless we make sure "?" is handled well, everywhere. That last part may be tricky to get right.

I think the problem here is that the error "cannot refer to unexported name types._" is at position 6:8 and the error "syntax error: unexpected const, expecting name" is at position 6:14. If the "first" error (which is actually later in time) were reported for 6:14, then it would probably not be shown because error reporting should be suppressing duplicate errors at the same position. That should be a much safer fix.

Also, consider how much effort should we put in to fix something that almost never happens?

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>