Update patch set 5
Patch Set 5:

> # go run run.go -- fixedbugs/issue11614.go
> issue11614.go:22: no match for `interface contains embedded non-interface int` in:
> 	issue11614.go:22: type interface { int } is not an expression

This is a real failure due to this CL, but it's a bit funny. It looks like the only reason we were printing "interface contains embedded non-interface int" here is because of printing the type for the other error message.

The easy workaround is probably to add "checkwidth(n.Type)" after the "yyerror("types %v is not an expression", n.Type)" in near the end of typecheck1.

Longer term, I wonder if we can get rid of all these random checkwidth calls by just putting one in the setTypeNode function you added.

(It would probably have to be guarded by "if !t.IsDDDArray() { ... }" since DDD arrays are weird.)

Patch-set: 5
Reviewer: Gerrit User 5440 <5440@62eb7196-b449-3ce5-99f1-c037f21e1705>