misc/lockcheck: check for lock cycles in all.bash
DO NOT SUBMIT. We have lock cycles right now, so this will fail.

This adds a dist test that runs the runtime and runtime/debug tests
under the runtime lock cycle detector.

Since this takes about 40 seconds, we only do it automatically on the
builders, which will be able to farm it out to run in parallel.

If the test detects a lock cycle, it dumps the whole lock graph to
stdout so it can be easily retrieved from the builder log and analyzed
separately.

Change-Id: I27716e7f50726211e42336bc47a8ac897da23623