Update patch set 1
Patch Set 1:

So then instead of two inconsistent ways there will be a third even more verbose way?

That is almost certainly not going to be approved.

Also, sending code isn't how we get new API into Go. It first needs to be proposed & accepted on the issue tracker.

I'm going close this for now.

Patch-set: 1
CC: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>