Update patch set 5
Patch Set 5: Run-TryBot+1

> Patch Set 4:
> 
> > os.Pipe() produces (*os.File)
> > 
> > and then the reading *is.File is wrapped in the io.LimitedReader to exercise the code path ways. https://golang.org/pkg/os/#Pipe
> > 
> 
> I am pretty sure that TransmitFile Windows API does not work with Windows pipes. See, for example, https://go-review.googlesource.com/c/go/+/79775/
> 
> Alex

Thank you for the reference Alex, that was quite informative, TIL! I've updated the test. If later we deem the test to be doing too much or be bulky we can remove it later, but for now I am using it on the trybots to validate my logic.

Patch-set: 5
Reviewer: Gerrit User 5137 <5137@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1