Update patch set 3
Patch Set 3:

If the regex already existed, perhaps clarify that in the commit message, like:

cmd/go: add a test for -sectcreate in LDFLAGS

It was already covered by a regex pattern, but it didn't have a test.

To fix the issue for good, add a regression test.

Fixes #28832.

Patch-set: 3
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>