Update patch set 2
Patch Set 2:

> If I interpreted your concern in issue correctly, if there's global symbols, the index of first global symbol will go in, otherwise, it's the number of local symbols, which what the code currently does.

Am I just not reading this correctly? I don't have an easy way to test this locally, but if s is nil, i will be 0, and if s is a global symbol, i will be 1, right? Presumably they should have the same behavior.

(I suspect the reason it's working when there's a global symbol has to do with the null symbol not being included in the list?)

Patch-set: 2