Update patch set 1
Patch Set 1:

> > I think that a better fix, that should work for all targets, would be to have `sigtrampgo` call `inVDSOPage`.  I haven't implemented that because I wasn't able to find a test case that would fail on the builders, perhaps due to #33574.
> 
> Can we really enter go functions without setting up the G register?
> This is actually the point I was unsure and tried to avoid.

We can if the functions are marked //go:nosplit and if the function does not produce any write barriers.  The latter will be enforced because sigtrampgo is marked //go:nowritebarrierrec.  It might be necessary to change inVDSOPage to use `for i := range vdsoSymbolKeys { k := &vdsoSymbolKeys[i]` to avoid generating a write barrier by copying a vdsoSymbolKey value.

Patch-set: 1