Update patch set 1
Patch Set 1:

> > > Do you have any idea on where to store G before going into VDSO functions?
> > 
> > Why do we have to?
> 
> I thought you meant that we would insert a branch in front of the body of sigtrampgo that conditionally restores the value of G.
> Could you clarify what you do plan to do after checking inVDSOPage?

If we could store the G value before entering a VDSO function in such a way that the signal handler can retrieve it, that would be nice.  But I don't see how to do that in general.

So what I'm suggesting is that if we are in the VDSO page on a GOARCH that stores the g value in a register, that we adjust the code to not require the g value.  Any code that looks at will need a g == nil variant.  I think this should be doable.

Patch-set: 1