Update patch set 6
Patch Set 6:

> Hello, this seems to break the plugin test on the FreeBSD ARM builder:
> 
>   ok  	path/filepath	0.449s
>   # plugin.test
>   /tmp/workdir-host-freebsd-arm-paulzhol/go/pkg/tool/freebsd_arm/link: running clang failed: 
>   exit status 1
>   clang: error: invalid linker name in argument '-fuse-ld=gold'
> 
> FAIL	plugin [build failed]
> 
> I'm not sure what adds the -fuse-ld=gold ldflag. The FreeBSD x86-64 and x86 builders don't seem to be effected.
> Perhaps I'm missing the llvm package which gets installed on the x86 images. It should contain the llvm60/lib/LLVMgold.so plugin.
> Should I update the ARM image?

The -fuse-ld=gold flag is most likely coming from (*Link).hostlink in cmd/link/internal/ld/lib.go, where it is used on ARM platforms when dynamically linking.  This CL is causing the plugin package to be tested on FreeBSD/ARM.  The test ought to pass, but it does cause the linker to select dynamic linking, and that triggers the -fuse-ld=gold case.

Probably we should change the linker to only add -fuse-ld=gold on GNU/Linux.  Want to see if that works and send a CL if it does?  I can write the CL if needed.  Thanks.

Patch-set: 6