Update patch set 1
Patch Set 1:

The current code fails because resetAll alters the state of the compressor, and so it matters when it gets called. Using reset affects when resetAll is called, and thus a problem.

So the intent of this change is something like "make it so that resetAll never alters the state of the compressor". At first glance, it looks like this isn't true - the first test here alters the state of the compressor by throwing away the table when prev is empty. If there can be relevant data in the table when prev is empty, and cur is high, then the state of the compressor gets altered and the bug comes back.

So how can prev be empty at a time other than initial startup? Anytime encode is called with "src" being empty. So now the bug happens depending on whether encode is ever called with an empty buffer. So now we need to carefully check the caller to be sure, but hey, it is true: encode is never called with an empty src. This still seems like a needlessly brittle fact, to make the correctness of the implementation depend on such a thing.

So it only helps in the rare case that resetAll is called directly by reset, and even then, it's not clear that it's any faster than going ahead and doing the table update step anyway.

If it is actually faster, then I think it would be better to do the table-clear as part of reset, rather than have a back-door dependency on exactly how encode is called.

Patch-set: 1
CC: Gerrit User 28777 <28777@62eb7196-b449-3ce5-99f1-c037f21e1705>