Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> The current code fails because resetAll alters the state of the compressor, and so it matters when it gets called. Using reset affects when resetAll is called, and thus a problem.
> 
> So the intent of this change is something like "make it so that resetAll never alters the state of the compressor". At first glance, it looks like this isn't true - the first test here alters the state of the compressor by throwing away the table when prev is empty. If there can be relevant data in the table when prev is empty, and cur is high, then the state of the compressor gets altered and the bug comes back.

There can't. When 'prev' is empty there is no history and everything in the table will point to invalid data. This is why adding "max match offset" is enough to make the entire table invalid.

Let me explain a bit more. This could maybe be added to the documentation as well.

The table contains matches. The index is the hash of the content.
Each entry in the table contains the value at that point in history.
The offset is the offset of the match with the current start offset (e.cur) added.
The current match offset specifies the offset at the start of the block while compressing.

The match offset is calculated as  relative_offset = current_offset - (candidate_offset - e.cur). If this offset is bigger than maximum possible offset we discard it.

Whenever history is added the length of the history is added to 'e.cur'. This means that when a new block starts the first byte of 'src' will have an offset of 'e.cur'. Indices in history will be negative after subtracting 'e.cur', so prev[len(prev) + offset - e.cur] will be the entry pointed to.

When a reset happens we add the maximum match offset to 'e.cur'. This means that all mathces will be rejected because we are beyond the maximum offset.

The table reset will change the e.cur offset to be max offset and adjust the offset of all valid offsets, so a table shift should be transparent in terms of compression, whereas it currently invalidates all matches and therefore affects compression.

> So how can prev be empty at a time other than initial startup? Anytime encode is called with "src" being empty. So now the bug happens depending on whether encode is ever called with an empty buffer. So now we need to carefully check the caller to be sure, but hey, it is true: encode is never called with an empty src. This still seems like a needlessly brittle fact, to make the correctness of the implementation depend on such a thing.

That is not an issue. If we don't have the history we can't use it for matching anyway and all matches will fail regardless of what e.cur is.


> If it is actually faster, then I think it would be better to do the table-clear as part of reset, rather than have a back-door dependency on exactly how encode is called.

That would have quite high impact on people using it for small payloads.

Do you have a test case, otherwise I will see if I can write one that executes in reasonable time, which should be possible with multiple resets - though it would be very implementation specific.

I do however think I found an unrelated issue when using "Sync" which could lead to a crash. I will investigate and open a separate issue.

Patch-set: 1
Reviewer: Gerrit User 6545 <6545@62eb7196-b449-3ce5-99f1-c037f21e1705>