Update patch set 1
Patch Set 1:

> > So how can prev be empty at a time other than initial startup? Anytime encode is called with "src" being empty. So now the bug happens depending on whether encode is ever called with an empty buffer. So now we need to carefully check the caller to be sure, but hey, it is true: encode is never called with an empty src. This still seems like a needlessly brittle fact, to make the correctness of the implementation depend on such a thing.
> 
> That is not an issue. If we don't have the history we can't use it for matching anyway and all matches will fail regardless of what e.cur is.

The bug is not a question of using the history for matching incorrectly or anything of the kind. It's about *discarding the table* in a way which depends on the history of the struct.

It's always safe to discard the table from a correctness perspective, the problem is that we must be sure to discard the table at exactly the same time for different runs.

In your patch, if someone calls encode with a zero-length src, and cur is big enough, you'll discard the table, and since cur is leaked across resets, this means that when the table is discarded depends on what happened before previous resets.

> > If it is actually faster, then I think it would be better to do the table-clear as part of reset, rather than have a back-door dependency on exactly how encode is called.
> 
> That would have quite high impact on people using it for small payloads.

Why are people with small payloads calling Reset?

Patch-set: 1