Update patch set 1
Patch Set 1:

> > > If it is actually faster, then I think it would be better to do the table-clear as part of reset, rather than have a back-door dependency on exactly how encode is called.
> > 
> > That would have quite high impact on people using it for small payloads.
> 
> Why are people with small payloads calling Reset?

Sorry, that's a silly question. They have plenty of good reasons for calling Reset.

But the question is, why do we care about Reset being super fast?

In other words, I don't think the table clearing in your first if is necessary, I think it might be harmful, and I think it's expensive to boot. (Or, if it's not expensive, then we can do it in every reset.)

Patch-set: 1