Update patch set 1
Patch Set 1:

> But the question is, why do we care about Reset being super fast?
> 
> In other words, I don't think the table clearing in your first if is necessary, I think it might be harmful, and I think it's expensive to boot. (Or, if it's not expensive, then we can do it in every reset.)

It is not super expensive, but will be significant if you compress a lot of 1KB payloads or less. The difference is writing 128KB of zeroes to memory or incrementing an int32 field.

Two things prevent the function from being called with a 0 size src:

1) On the only caller, there is a check for sizes < 128 bytes.
2) The `if len(src) < minNonLiteralBlockSize` in the encode function will invalidate history.

Patch-set: 1
Reviewer: Gerrit User 6545 <6545@62eb7196-b449-3ce5-99f1-c037f21e1705>