Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > But the question is, why do we care about Reset being super fast?
> > 
> > In other words, I don't think the table clearing in your first if is necessary, I think it might be harmful, and I think it's expensive to boot. (Or, if it's not expensive, then we can do it in every reset.)
> 
> It is not super expensive, but will be significant if you compress a lot of 1KB payloads or less. The difference is writing 128KB of zeroes to memory or incrementing an int32 field.
> 
> Two things prevent the function from being called with a 0 size src:
> 
> 1) On the only caller, there is a check for sizes < 128 bytes.

I know, but that's brittle. It's otherwise not an error to call with empty src.

> 2) The `if len(src) < minNonLiteralBlockSize` in the encode function will invalidate history.

In fact, thanks for pointing me at that. prev can be set to empty in that case simply with a small non-zero src.

The point isn't whether it invalidates history, again. The point is that the behavior of the compressor will depend on what happened before reset.

Patch-set: 1