net: permit sendFile on more than *io.LimitedReader on Windows
CL 192518 was a minimal simplification to get sendfile
on Windows to work with chunked files, but as I had mentioned,
it could use some simplification.

This CL improves it by:
* Removing the decrement of lr.N (the *io.LimitedReader), on
each invocation of doSendFile
* If the reader is not an *io.LimitedReader, since the underlying
reader is anyways an *os.File, we fallback and stat that
file to determine the file size and then also invoke the chunked
sendFile on the underlying reader. This issue existed even
before the prior CL.
* Adding stateful helpers to track the remaining bytes.

Thus if the callers of sendFile don't send an *io.LimitedReader,
but have a huge file, we can still invoke the chunked sendFile
on it.

Change-Id: I97a67c712d558c84ced716d8df98b040cd7ed7f7