internal/poll, net: permit sendFile on more than *io.LimitedReader on Windows
CL 192518 was a minimal simplification to get sendfile
on Windows to work with chunked files, but as I had mentioned,
I would add even more improvements.

This CL improves it by:
* If the reader is not an *io.LimitedReader, since the underlying
reader is anyways an *os.File, we fallback and stat that
file to determine the file size and then also invoke the chunked
sendFile on the underlying reader. This issue existed even
before the prior CL.
* Extracting the chunked TransmitFile logic and moving it directly
into internal/poll.SendFile.

Thus if the callers of net.sendFile don't send an *io.LimitedReader,
but have a huge file, we can still invoke the chunked
internal/poll.SendFile on it and now even direct invokers can send
files >2GiB.

Updates #33193.

Change-Id: I97a67c712d558c84ced716d8df98b040cd7ed7f7