Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Not sure I agree. There are dozens of mentions of []byte across std docs, and I don't think we should repeat this everywhere.

I was also not exactly sure if this was a good idea when i made this pr but because i've had a really hard time debugging a problem i've submitted this pr.  
But now that i'm thinking of it the people that use uint8 slices probably already know that it's an alias of []byte.

Patch-set: 2
Reviewer: Gerrit User 33468 <33468@62eb7196-b449-3ce5-99f1-c037f21e1705>