Update patch set 3
Patch Set 3:

> This is a behavior change, right? Aren't we concerned about it breaking code in the wild? Should this go through proposal review?

Please see the original issue thread. I interpret the docs to agree with the behavior here:

> Otherwise Unmarshal reuses the existing map, keeping existing entries.

This is the same point that @as brings up in his original post.

Certainly, some tests or programs could break, which is why I want to attempt merging this early in the cicle.

If you think a proposal is needed, we can turn the issue into one and wait for the committee. Though then we're probably late to try this for 1.14. I made a decision as some time had passed, it seemed clear to me, and the other maintainers hadn't chimed in (though I never pinged them directly, that's my bad).

Patch-set: 3
Reviewer: Gerrit User 13550 <13550@62eb7196-b449-3ce5-99f1-c037f21e1705>