Update patch set 6
Patch Set 6: Run-TryBot+1

> Patch Set 5: Code-Review-1
> 
> > Patch Set 5:
> > 
> > TryBots beginning. Status page: https://farmer.golang.org/try?commit=8575866d
> 
> This CL fails non-short tests for me. Still investigating the reason.

Oddly, the test passed when I ran it a second time.
I suspect that the failure may be related to #34092, which is an existing issue (and if so does not need to hold up this CL).

Patch-set: 6
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1