encoding/json: This CL adds Decoder.InternKeys
The existing decoder creates a new string for every key.
For large JSON input where keys are repeated, we can avoid
this by interning keys. This change adds a Decoder.InternKeys
function that causes the decoder to store keys in a
map[string]string and uses a single copy for repeated keys.
The default behavior is not to intern keys.

This change only affects Decoder when decoding to an interface{}
or, when tokenizing the input. It does not affect Unmarshal, or
decoding to a struct.

Fixes #32779

Change-Id: I8a0557283e67d144a6f2bc5c04ef9a63fc9c7dcf
GitHub-Last-Rev: d07e7fd3ba017a937e5412e20415ab0aa32b6bf4
GitHub-Pull-Request: golang/go#33416