[dev.link] cmd/internal/goobj: support parsing new object file
DO NOT REVIEW

Add support of parsing new object file format. We use the new
parser if the magic string matches the new one, otherwise use the
old one.

The parsed data are still filled into the current goobj API. In
the future we may consider to change the goobj API to a close
match of the object file data.

Now objdump and nm commands support new object file format.

Change-Id: I06d05b2ca834ba36980da3c5d76aee16c3b0a483