runtime: add option to scavenge with lock held throughout
This change adds a "locked" parameter to scavenge() and scavengeone()
which allows these methods to be run with the heap lock acquired, and
synchronously with respect to others which acquire the heap lock.

This mode is necessary for both heap-growth scavenging (multiple
asynchornous scavengers here could be problematic) and
debug.FreeOSMemory.

Change-Id: I24eea8e40f971760999c980981893676b4c9b666