Update patch set 1
Patch Set 1:

Was this new API discussed anywhere? In general, Gerrit is not the right place to discuss those, so a proposal issue would be best.

From what I recall on the previous discussions of json's use of sync.Pool, that issue was possible to fix without new API, no?

Also, why is this an "updates" and not a "fixes"?

Patch-set: 1