Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Was this new API discussed anywhere? In general, Gerrit is not the right place to discuss those, so a proposal issue would be best.
> 
> From what I recall on the previous discussions of json's use of sync.Pool, that issue was possible to fix without new API, no?
> 
> Also, why is this an "updates" and not a "fixes"?

If this needs more discussion first, we can close this and discuss on GH.

The last proposal on the issue was a new API (although I must admit, it's not clear to me when a comment on GH constitutes something that should be implemented, and when it's just a random idea awaiting further discussion. After 18 days of no further activity, I decided to try tackling it.) This isn't quite as extensive as the new API proposed, which is why it's 'updates' and not 'fixes'--this may not be a complete fix.

Patch-set: 1
Reviewer: Gerrit User 33388 <33388@62eb7196-b449-3ce5-99f1-c037f21e1705>