Update patch set 1
Patch Set 1:

Ah I see, thanks for clarifying. I was confused because the Gerrit notification arrived before the one from your github comment.

I think having this CL around is fine, but perhaps mark it as a WIP to clarify that we aren't yet sure if it's the right solution.

The commit message should also explain why this is useful, and why it updates (and not fixes) the issue. But no need to do that while the CL is still a WIP :)

Patch-set: 1